Error handling bug fix in HEMCO initialization
If the scale factor ID was not found in HEMCO_Config.rc during
initialization then HCO_ERROR was called to print a message to the HEMCO
log file. However, the HEMCO log file does not exist at this point in
the run so no message was printed, making debugging scale factor config
problems difficult. This update replaces the call to HCO_ERROR with
prints to std out.

Also in this commit, error messages are mpw printed out indicating what
part of HCO_INIT or SetReadList failed. Previously if RC not equal to
HCO_SUCCESS the subroutines would return without message.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>