Fix error checks for switches and HEMCO_Config.rc instead of overwriting those settings
Several settings in HEMCO_Config.rc were getting overwritten based on settings
in input.geos. This can cause confusion to users who don't realize this is
happening unless they dig in the log files. Instead, we now compare the
settings in HEMCO_Config.rc against the settings in input.geos and stop the
simulation with an error message if there's an inconsistency.

Also added messages to the log file when emissions or chemistry are turned
off in input.geos indicating that the input data for that component will
not be read.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>