Bug fix: allow turning off emissions in GCHP
HEMCO now processes non-emissions data as well as emissions inputs.
Therefore HEMCO must be executed at a high-level even if emissions
are turned off in input.geos. This is the case in main.F for
GEOS-Chem Classic but not gigc_chunk_mod.F90 in GCHP, causing a
run error in GCHP if emissions are turned off in input.geos.
Removing the criteria fixes the issue.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>