Fixed bug in ND51 and ND51b when testing for min/max values of I and J
We have removed the I0 and J0 values (which are now defined by FlexGrid)
from the error traps where we check to see if ND51_IMIN & ND51_IMAX are
valid and if ND51_JMIN & ND51_JMAX are valid.  Ditto for the ND51b
diagnostic.

After FlexGrid was implemented, we did not test with the bpch
diagnostics, because these diagnostics are deprecated.  We look
to the GC community to report issues such as these with the remaining
bpch diagnostics.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>