Remove commented out grid comp name check in diaglist_mod
The commented out code should never be necessary in GCHP since elsewhere
in the file there is a check for grid comp name GIGCchem. Lines without it
are skipped when adding objects to diaglist. Non-GIGCchem grid comp
entries in HISTORY.rc are handled separately by MAPL and are not
registered in GEOS-Chem. Their presence should not stop the run.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>