Bug fix: Remove hardwired timestep for TOMAS simulations
Dana McGuffin reported that the ADT variable in AEROPHYS was hard-coded
to 3600 but this should really be defined as ADT = GET_TS_CHEM().

Also: the ADT variable does not need to be held !$OMP PRIVATE for the
parallelization, as its value will be the same across all threads.
This will prevent calling GET_TS_CHEM for each (I,J,L) iteration.

We have also bracketed the AD61_INST with an IF statement to make sure
that this array is not accessed unless ND61 diagnostics are turned on.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>