GCHP rundir updates: auto-update KppDiags diag collection freq/dur/mode
Previously the KppDiags diagnostic always output at hourly frequency in
GCHP unless manually changed in HISTORY.rc due to an omission in
runConfig.sh.

This commit also includes an update in HEMCO_Config.rc that does not
impact GCHP, for purposes of keeping the GCHP and GCC config files in
sync.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>