Use new MEGAN file with time unit attribute fix to avoid GCHP gfortran bug
MAPL requires the hour/min/day string in the time unit in netcdf files
to be all integers, e.g. 00:00:00, but the old file had 0.0 for seconds.

Also in this commit, uncomment CMIP6_Sfc_CH4 in ExtData.rc so that it is
ready to use if turned on in HEMCO_Config.rc.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>