Alphabetize species in GCHP input files
With the increasing number of species in the full-chemistry simulations, it's
becoming increasingly difficult to compare the list of species when they are
updated. To simplify things, we will list the species in alphabetical order in
the Advected Species Menu of input.geos and in HISTORY.rc. This also allows
for easier comparison with the species database, netCDF output files, etc.
where species are listed in alphabetical order as well.

Also fixed the order of some species in the KPP .eqn files.

Finally, updated GEOS HEMCO_Config.rc file to be consistent with recent
changes.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>