Switch GCHP wind regridding to bilinear interpolation yet again
This repository is used in both GCHP and in new GCHP wrapper called
GCHPctm. Conservatively regridding winds works in the latter but not in
the former given how the inputs are configured in ExtData.rc and the
version of MAPL used. Until retiring GCHP and switching to GCHPctm, which
will happen in 13.0.0, we will keep the default bilinear interpolation.
A separate branch of GEOS-Chem will be created with conservative
regridding of winds for use in GCHPctm until it becomes standard.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>