Write the first interface routine for abstracting init/registration
We have added several updates to state_diag_mod.  We are going
to add several module routines overloaded by the "Initialize" interface
in order to abstract the process of getting a mapping array,
allocating the array, and registering a diagnostic.  This should
make adding new diagnostics much easier.

NOTE: Some debug output is left in this commit, we will clean it
up later.  We wanted to stash the results here before moving forward.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>