Finish adding Init_and_Register routines; abstract and reuse code
In state_diag_mod.F90, we have renamed the Initialize_* routines
to Init_and_Register_*.  These routines are overloaded by an interface
so that they can accept different array sizes and types (e.g. REAL*4
2D/3D/4D, and REAL*8 2D/3D/4D).

We have also abstracted code from the Init_and_Register_* and
the Register_Diag_* subroutines into reusable subroutines in order
to minimize code duplication.  This helps to reduce the module size.

NOTE: There is still a debug STOP in this commit, but we will remove
it later.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>