Only perform pole-averaging of regridded data if actually regridding to the poles.
This commit fixes an issue affecting GEOS-Chem Nested, MODEL_WRF, MODEL_CESM (in HEMCO_CESM) in which Regrid_A2A performs northernmost and southernmost quantity averaging regardless of whether the boundaries are actually on the poles.

Now the check is actually performed, because REGRID_A2A is being increasingly used for regional purposes.

Signed-off-by: Haipeng Lin <hplin@seas.harvard.edu>