Add initial version of species_database.yml based on GEOS-Chem 12.8.0
The species database file added here was obtained from  the dev/0.2.0
branch in gcpy. It was created by converting the species_database.json file
saved out in the 1-month GEOS-Chem 12.8.0 benchmark to YAML in python.

The next steps will be to add all GEOS-Chem species to this file, to define
molecular weights for all species, and general cleanup.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>