Use GCHP heartbeat for GIGCchem alarm reference time
This update makes the behavior of the GIGCchem alarm consistent across all
timesteps a user might configure in runConfig.sh. The reference time is
now automatically set in GCHP.rc (via runConfig.sh) to the heartbeat in
HHMMSS. This is necessary for the GIGCchem alarm to always be true the
first timestep and every other timestep following.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>