Revert HEMCO to prior new repo sync, but keep subdirectories in src folder
Using the new HEMCO repository that will be part of GEOS-Chem 13.0
requires extensive changes to the GEOS-Chem Classic legacy build system
that will be retired in 13.0 after the move to the new GEOS-Chem Classic
repository wrapper. We are therefore keeping the simpler legacy HEMCO
build system in GEOS-Chem Classic until 13.0. Otherwise HEMCO in this
repo and the HEMCO repository are identical.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>