Add species for specialty simulations to species_database.yml
Anchors have been introduced to avoid repeating the same information over
and over in the species database. This is especially useful for tagged
simulations, where the species differ only in their name.

Separate species species database yaml files have been added for the APM
and TOMAS simulations to avoid clutter in the standard file. In GEOS-Chem
we should only read the APM and TOMAS files when the APM and TOMAS compile
options are enabled.

Also made the following changes to species_database.yml:
- Changed Is_Aero to Is_Aerosol for clarity
- Added missing BackgroundVV definitions for several species