Bug fix: Restore a missing Container%FirstInst=FALSE statement
The commit 28109a7, as described in this Github issue:

   https://github.com/geoschem/geos-chem/issues/269

originally fixed an error where instantaneous files having data at 00 UTC
were being overwritten by the next day's output.  However, in commit
28109a7, we moved the statement Container%FirstInt=.FALSE lower in the
subroutine so that it could be used again.  However, we forgot to add
this line in the IF block that returns out of the subroutine.

Therefore, due to this bug, any existing instantaneous file would get
overwritten inadvertently.  This new bug fix now prevents this, as well
as setting the netCDF reference time properly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>