Nullify the State_Chm%TLSTT variable in Init_State_Chm
The State_Chm%TLSTT variable is now initialized to NULL in the routine
Init_State_Chm.  Prior to this commit, State_Chm%TLSTT was being
initialized with junk values, which could have given the impression
that it was allocated.  Then this would result in a seg fault when
Cleanup_State_Chm tried to deallocate this.

This error was made evident in specialty simulations where Linoz was
not being used.  This commit should now fix this error going forward.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>