Updates for anchors in YAML files and for defining SpcData from YAML
In a YAML file, the << characters should be used when pointing to
a YAML anchor.  We have replaced the >> characters in the various
species database files with <<.  Also, we have made sure that each
anchor appears higher up in the file than all variables that point to
it.

We also have ensured that the species_database*yml files can be
parsed properly in to Python.

Also, we have corrected some issues that occured in transferring the
data from species database YAML files to the State_Chm%SpcData object.
Namely, we rewrote some logic that was overwriting the WD_KcScaleFac
values, and we make sure to round off all real values to 2 decimal
places.

We now use qfyaml_mod.F90 from QFYAML 0.2.1.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>