Avoid memory corruption bug in HEMCO when certain debug flags are on
This issue was originally reported by Sebastian Eastham (MIT). He wrote:
"It seems that HCO_EvalFld is being called with HcoState%Grid%PBLHEIGHT%Val,
but that array is still null (as set during the state initialization).
Since HCO_EvalFld expects an already-allocated array, this results in an
error when certain debug flags are on, and presumably some kind of memory
corruption when they're not."

This issue is fixed by calling HCO_ArrAssert prior to HCO_EvalFld. The
assert function will allocate the array to the proper size if it is not
already allocated.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>