Fix a few more issues and add sanity checks
species_database.yml: Fixed a couple more typos (OCPO: WD_AerScavEff,
PAN: WD_RetFactor) so that we match the prior codde.

species_database_mod.F90: Added more sanity checks so that we zero
aerosol fields for gas-phase species and vice-versa.  This is necessary
to match the behavior of the prior code.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>