Add species database fixes for CH4, CO and Hg species
CH4 species: Remove the CH4:Background_VV tag from CH4_PROP and add it
to the CH4 species.  This was causing the CH4:Background_VV tag to be
applied to all of the tagged CH4 species.

TagCO species: Change the species names in species_database.yml
to match those in the run directory input.geos (i.e. case-sensitive
instead of all-caps).

Hg species: Update the ThisSpc%Hg_Cat field in species_database_mod.F90,
which is necessary to properly index the emissions for Hg simulations.