Fix incorrect indexing of aerosol species undergoing hygroscopic growth in aerosol_mod.F90
The State_Diag%*Hyg* arrays were mixed up following the alphabetizing of
the species in the Advected Species Menu in input.geos. Before this update,
the order of species was 1=SO4, 2=BC, 3=OC, 4=SALA, and 5=SALC in both the
arrays from RD_AOD and in RDAER and in the State_Chm%MapHygGrth array used
map species to diagnostics using the ?HYG? wildcard in HISTORY.rc. Following
the alphabetizing of species, the array in State_Chm%MapHygGrth was changed
to 1=BC, 2=OC, 3=SALA, 4=SALC, 5=SO4. To make sure the correct indexing, is
used, we now consistently use the ordering 1=BC,...,5=SO4 when using the
State_Diag%*Hyg* arrays and 1=SO4,...,5=SALC in reference to arrays from
RD_AOD and RDAER (by using the MAP_NRHAER array in aerosol_mod.F90).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>