Divide loop timers by the number of threads
For timers within parallel loops, each timed task is being applied on N
threads.  If each task takes T second, then the timers would return
a total time of N*T seconds, which may be misrepresentative in the
timing tables.

A better representation is to divide the total time of a loop timer
by N threads.  This gives an "average" time per task across all cores
but is more in line with the total time measured by a timer that
straddles the parallel loop.

Another thing we can do is to try to measure the amount of "real" time
instead of CPU time that each task takes (i.e. use CPU_TIME function
in Timer_TheTime.  That might obviate the need to divide by the number
of threads.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>