Move MODEL_WRF return to top of INIT_DRYDEP.
This change for MODEL_WRF fixes the issue that an allocation check and
return was not early enough, resulting in a second function call
zeroing out internal species indices and breaking dry deposition instead.

NOTE: This update was originally made by Haipeng Lin for 12.9.0, but
was manually applied by Bob Yantosca for 12.8.2, in order to avoid
pulling in other 12.9.0 modifications.

Signed-off-by: Haipeng Lin <hplin@seas.harvard.edu>
Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>