Add option to set path to species_database.yml in input.geos
Previously, the species database was assumed to be located in
./CodeDir/Headers/species_database.yml. However, this was causing issues
when CodeDir wasn't defined, as is the case in difference tests and unit
tests. We now instead specify the path to that file in the Simulation
Menu of input.geos.

Also removed code to separately read the APM and TOMAS species database
files. Instead, code in the run directory creation scripts has been
modified to append the contents of those files to species_database.yml to
have a single input file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>