Dryrun fix: Do not attempt to go back in time if a file is not found
We have placed a shunt in routine SrcFile_Parse (in module
HEMCO/src/Core/hcoio_read_std_mod.F90) for the GEOS-Chem/HEMCO dry-run.

Prior to this fix, if a file (say for 2019/08/10) was not found,
the program flow would enter a loop and attempt to keep going back
in time until the most recent file was found (say 2019/08/01).  This
resulted in GEOS-Chem/HEMCO dry-run simulations not being able to
pick up missing files if at least one file already existed.

The fix is to exit routine SrcFile_Parse before entering the loop
where we attempt to go back in time if we are doing a dry-run
simulation.

We will make a corresponding pull request to the HEMCO repository.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>