Do not read FLASH_DENS and CONV_DEPTH fields in FlexGrid when lightning NOx extensionin HEMCO is off
A shadow variable Input_Opt%DoLightNOx has been added and will be set to
true or false depending on if the lightning NOx extension is turned on in
HEMCO_Config.rc. When that option is off, the reading of FLASH_DENS and
CONV_DEPTH in flexgrid_read_mod.F90 will be skipped (keeping those fields
as all zeroes). This will avoid the requirement of having users download
those data for simulations that do not require it.

For more details, see https://github.com/geoschem/geos-chem/issues/279.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>