Fix bug in RRTMG with stratospheric aerosol
Two of the logical checks in the RRTMG interface were incorrectly using the variable `IS` outside of the loop in which it was defined. This resulted in the stratospheric aerosol asymmetry parameter and single scattering albedo being incorrectly specified - they had been summed correctly inside the loop, but were not being divided by total AOD outside the loop. Removing these two logical checks resolves the issue. Now instead the code checks for the presence of any aerosol, which will be slightly slower but much safer.

This resolves issue #346 - applying this patch results in a negative direct radiative effect from stratospheric aerosol, which is expected.