Add option to get offline HCl fields to ISORROPIA
Following the halogen chemistry updates, ISORROPIA expects HCl as a
species; however HCl is not included in the aerosol simulation since it is
a gas-phase species, so that simulation crashes. In isorropiaII_mod.F90
other gas-phase species like HNO3 utilize code to check if it's a full-
chemistry or aerosol-only simulation. If it's a full chemistry simulation,
ISORROPIA takes the online species concentration from State_Chm%Species. If
it's an aerosol-only simulation, ISORROPIA gets offline fields read in via
HEMCO. Updates to do the same for HCl have been added here, however we do
not currently have HCl fields archived for use as input in aerosol-only
simulations. We will save out these fields as part of the 1-year benchmark
for 12.9.0 and update the HEMCO_Config.rc file for aerosol-only simulations
in 13.0.0 to use those archived fields.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>