Define State_Chm%HetInfo to hold hetchem species metadata
We have added the State_Chm%HetInfo sub-field, which is an object that
holds the modelId, MW_G, Henry_K0, Henry_CR fields of species that
are used in heterogeneous chemistry.  This is necessary so that we
can pre-compute these metadata before hetchem is called, which should
be much more computationally efficient.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>