Fix unit conversion error in gosat_ch4_mod.F
The unit conversions in CALC_GOS_CH4_FORCE were still assuming units coming
into the routine were in kg/box. This is no longer the case. The units of
State_Chm%Species from main.F are in kg/kg and need to be converted to v/v.
This is now done via a call to Convert_Spc_Units before and after the loop
over grid boxes.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>