Update TaggedDiagItem%TagList indexing to correct negative values
The indexing algorithm for individual items in TaggedDiagItem%TagList
was updated in order to remove the possibility of having negative
index values.  Instead, each tag is numbered sequentially.

NOTE: A further commit will be needed to fix the UVFlux* diagnostics,
in which the tagname lookup depends on a specific index ordering.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>