Implement proper indexing for DryDepMix in Compute_Sflx_for_Vdiff
In the routine Compute_Sflx_for_Vdiff (located in module
GeosCore/hco_utilities_gc_mod.F90), we have fixed an error in
updating the State_Diag%DryDepMix diagnostic field.

This array has to be indexed with (I,J,S), where S is the diagnostic
slot ID obtained from State_Diag%Map_DryDepMix%id2slot.  But it was
being indexed with the drydep index ND.  This can lead to an
out-of-bounds error if the diagnostic is allocated with less than
State_Chm%nDryDep slots.  Now fixed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>