Initialize RC and FOUND arguments in HEMCO routine SrcFile_Parse
The RC and FOUND arguments should have been initialized to default
values but weren't.  On some compilers, this means that they can get
set to junk values which may flip the value of the arguments to
unintended values.

For example, FOUND was being set to TRUE, which caused the GEOS-Chem
dry-run simulations to treat missing files as if they were present on
the system.  This would cause those files not to be downloaded.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>