Change RCP folder in run/GCHPctm from RCP/v2015-02 to RCP/v2020-07
This commit changes the folder from which RCP emissions are read
from disk by HEMCO.  This corresponds to Github issue:
https://github.com/geoschem/geos-chem/issues/369.

NOTE: RCP emissions are an optional dataset.  They are turned off by
default, so this update will not affect the standard or benchmark
simulations unless a user actively turns them on in HEMCO_Config.rc
and in ExtData.rc.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>