Bug fix: CO2 simulation tagged species need to be in mixed-case
Monika Szelag (@monszel) pointed out that the tagged CO2 species names
were all in uppercase in Headers/species_database.yml, but that this
is inconsistent with usage in HEMCO_Config.rc and input.geos.

We have now changed the species database tagged CO2 species names
so that they are in mixed case (e.g. "CO2ff" instead of "CO2FF").

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>