Bug fix for GEOS-Chem dry-run: set Found = HasFile in SrcFile_Parse
In HEMCO routine SrcFile_Parse, there is an error check for the
GEOS-CHem dry run that prevents searching backwards in time to find
missing file.  But this error check was not setting the FOUND argument
to the proper value.  We have now fixed this.

Thanks to Lee Murray (@ltmurray) for the solution.

This commit resolves the issue reported in:
https://github.com/geoschem/geos-chem/issues/379

We will also make the corresponding commit in the HEMCO repository.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>