Update .gitignore files in GEOS-Chem source code folders
The .gitignore file in each folder tells Git to ignore certain types
of files based on their filename or extension (including the wild card
character "*").  We have updated the .gitignore files in most GEOS-Chem
folders to also ignore core.* files.  A few folders did not have a
.gitignore file, so we added them.

This will reduce the chance that large file such as corefiles or
executables will be added to the Git repository.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>