Bug fix: Make sure INIT_PRESSURE is called for the dry-run simulation
In routine GC_Init_Extra, the routine INIT_PRESSURE (which computes
the pressures on the vertical grid) was never being called.  The dry-run
program flow exited right after the INIT_DRYDEP routine was called.
But because INIT_PRESSURE was located further down in the module,
this caused a segmentation fault due to an out-of-bounds error.

The simple fix is to move the INIT_PRESSURE to be the first of the
routines called by GC_INIT_EXTRA.  Also we exit immediately after the
call to INIT_DRYDEP, whether or not Input_Opt%LDRYD=T.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>