HEMCO bug fix: override IsModelLevels if data provided is on sigma levels
We discovered that HEMCO was always interpreting data with vertical levels
as if they were on GEOS-Chem levels.  The fix is to override this
assumption if the lev:long_name attribute contains the text
"atmospheric_hybrid_pressure_sigma_levels".  In that case, HEMCO will
use the hyam and hybm parameters to place the emissions properly.

This commit addresses the issue described in Github issue:
https://github.com/geoschem/geos-chem/issues/391

A corresponding fix will be made in the standalone HEMCO repository.

This fix is being placed into GEOS-Chem 12.9.3 (we have updated the
version number accordingly).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>