Bug fix: Initially zero Item%nUpdates to avoid compiler side-effects
In History/histitem_mod.F90, we now zero the Item%nUpdates field to
zero.  Apparently this value was never given an initial value in
Init_HistItem.

On some compilers (e.g. gfortran 9.3 and later), uninitialized values
are not always guaranteed to be set to zero.  Explicitly defining
Item%nUpdates = 0 will help to make the code more robust and less prone
to unintentional side-effects.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>