Bug fix: always allocate State_Met%IMIX and State_Met%FPBL
In Init_State_Met (in Headers/state_met_mod.F90), we had an IF block
that only allocated State_Met%IMIX and State_Met%FPBL if PBL mixing
was turned on.  This can lead to a segmentation fault if PBL mixing is
turned off.  The most robust fix is to always allocate these fields.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>