Initial updates for the MeanOH diagnostic collection
We have added diagnostic arrays to the State_Diag object (in
Headers/state_diag_mod.F90) for computing the mass-weighted mean CH4
and OH concentration per full-atm or trop-only columns.  This is
necessary in order to bring this diagnostic into GCHP.  (In GEOS-Chem
classic, this is printed to the end of the log file but this is not
possible when we run GEOS-Chem with distributed parallelization.)

In GeosCore/flexgrid_mod.F90, we have also added a new routine called
Diag_Mean_OH_and_CH4, which computes the mean OH in the same way
as was done in the old diag_oh_mod.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>