Rename KPP mechanism from Standard to 'fullchem'
The KPP full-chemistry mechanism has been renamed to fullchem. The use
of 'standard' is no longer needed since there is one full-chemistry
mechanism. That mechanism may be applied to the troposphere+stratosphere
to reproduce UCX-based mechanisms or to the troposphere-only to reproduce
tropchem mechanisms in earlier versions. When using the troposphere-only,
linearized prod/loss rates archived from an earlier UCX-based run will
be read in via HEMCO and applied above the tropopause.

The fullchem mechanism has been rebuilt here to include updates to the
.eqn file made in the previous commit as well as file path updates made
here.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>