Update RRTMG output names to enable binary diagnostic output
The recent RRTMG updates to diagnostics changed the name of the species
outputs, e.g. O3 is now NOO3 for ozone. However, the names did not change
when setting Input_Opt%LSpecRadMenu, used for determining what binary
diagnostics to output. This resulted in only outputting binary RRTMG
diagnostics for the Baseline output, always on by default.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>