Update default background concentration for CH4 in species_database.yml
Background_VV was previously set to 1700 ppb for CH4 in the species database.
This is way to low compared to present-day values. The default value has
now been updated to 1800 ppb at the recommendation of Daniel Jacob.

Duplicate species database files and the test_yml.py script have been removed.
The species database files for TOMAS and APM can now be found in run/shared/.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>