Scale RRTMG diagnostics in first radiative timestep to correct netcdf diag
Without this scaling the RRTMG time-averaged netcdf diagnostics
incorrectly weight the first output with zeros since the state_diag
arrays are not updated until the mid-point of the RRTMG timestep. This
update scales the first values by 1.5 to cover the initial period leading
up to the first RRTMG computation.

Note that this will result in incorrect first values in RRTMG
instantaneous netcdf diagnostics since the values will be scaled by 1.5.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>