Bug fix for DUSTBIN wild card in reduction of diagnostic memory
In routine Get_Mapping (in Headers/state_diag_mod.F90), the isDustBin
logical should have been added to the definition of skipInd (which
indicates that we should not use the Ind_() function when we define
the mapData%slot2Id).  This omission had caused an error when certain
fields of the Aerosols mass were requested in HISTORY output.  This
has now been fixed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>