Change units of Metrics collection fields to fit within REAL*4 precision
The Metrics collection fields that had units of molecules or molec/cm3
have now been converted to units of kg or kg/m3.  This is necessary to
keep values smaller ~ 1e38, which is the maximum number representable
in REAL*4 precision.

The metrics_fullchem.py script that ships with the run directories has
been updated accordingly.

Also, the "MeanOHcolumn* and "MeanCH4column*" fields of the Metrics
collection have been replaced with "OHwgtByAirMass*" and
"CH4wgtByAirMass*" to more accurately denote that they are OH and CH4
concentrations multiplied by air mass.

Lastly, we have added whitespace changes to physconstants.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>