Allocate but do not register DryDepChm, DryDepMix if not explicitely requested
The State_Diag%DryDepChm (drydep in chemistry) and State_Diag%DryDepMix
(drydep in mixing) fields are used to compute State_Diag%Drydep
(overall dry deposition).  However, this required that DryDepMix,
DryDepChm, and DryDep be explicitly listed in the HISTORY.rc
diagnostics configuration file.

We have made a modification to Headers/state_diag_mod.F90 such that
the arrays (and mapping objects) for the DryDepChm and DryDepMix
diagnostics will be defined and set to that of the DryDep diagnostic,
without registering the diagnostics.  This will prevent DryDepChm
and DryDepMix from being archived to netCDF if only the overall
DryDep diagnostic is desired.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>