Remove DryDepChm and DryDepMix collections from HISTORY.rc templates
The DryDepChm and DryDepMix collections are no longer needed for
computing the total DryDep collection.  The arrays for DryDepChm
and DryDepMix will now be allocated regardless of whether or not
DryDepChm and DryDepMix are found in the HISTORY.rc file.

Therefore, to save disk space, we have removed DryDepChm and
DryDepMix fields from the template HISTORY.rc files.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>