Update date range for QFED in HEMCO_Config.rc to extend to 2020
QFED data through July 2020 are now available in HEMCO/QFED/v2018-07. The
date range in HEMCO_Config.rc files has been updated accordingly. The
time cycle flag for QFED has also been changed to EF to force HEMCO to stop
with an error when trying to use QFED emissions for a date that is not
available.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>