GCHP updates for mass tuning factor handling in HEMCO_Config.rc
Mass tuning factor is only relevant for benchmark smulations and previous
updates were made to only auto-update it if online dust emissions are on
(only occurs in benchmark by default). However, this means the keyword
@MASS_TUNING_FACTOR remains in HEMCO_Config.rc. That keyword is no
longer actually used, but it is referenced in comments only.

This commit replaces @MASS_TUNING_FACTOR with a default value (for c48)
and updates runConfig.sh to tell users to add a new value in a condition
for their non-standard target resolution there if not found. Users do
not need to comment out anything in runConfig.sh as was previously
instructures. Users also do not need to manually edit HEMCO_Config.rc.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>