Turn off GCHP periodic checkpoint output by default; add irregular option
GCHP will always output a checkpoint file at the end of the run but also
has the capability of doing programmable checkpoints. Previously runConfig.sh
allowed setting a regular frequency to output checkpoints. The reference
datetime was always set to the simulation start datetime. The RECORD_*
fields in GCHP.rc were always left uncommented, meaning periodic
checkpoints were always on. If the frequency was set to 000000 then it
was auto-updated to a large number so that no periodic checkpoints would
be written; however, MAPL still wrote an initial checkpoint in the first
timestep with this setting, confusingly filled with all zeros due to the
calling order in initialization.

This commit improves the usability of periodic checkpoints in GCHP.
They can now be turned entirely off from runConfig.sh, bypassing the
costly writing of an initial checkpoint filled with zeros. Users can also
turn periodic checkpoints on from runConfig.sh and configure params to
do either regular or irregular checkpoints in time. If outputting regular
checkpoints then the reference time can be auto-updated to be simulation
start time or users can specify a custom reference time.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>