Restore EmisNO_Fert diagnostic in HEMCO_Diagn.rc for full-chemistry simulations
This fix requires and update to hcox_soil_nox_mod.F90 in HEMCO to create
and link fertilizer NOx emissions to the correct field in that module.
Saving out EmisNO_Fert is only possible when the online SoilNOx extension
is used (i.e. only soil+fertilizer NOx emissions are available with the
offline soil NOx emissions).

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>