Fix registration issues in state_met_mod.F90 (and general cleanup)
We have restructured the initialization and registration process
in state_met_mod.F90 to match that of state_chm_mod.F90 and
state_diag_mod.F90.  An overloaded interface (Init_and_Register) is
called to allocate each State_Met field and to register it for HISTORY
diagnostics.

We also corrected an outstanding issue that fields with a dimension
of NSURFTYPE (i.e. # of Olson land types) were being registered as
a single (X,Y,Z) array instead of NSURFTYPE (X,Y) arrays.  This is
necessary to save these fields to HISTORY diagnostics, if so desired.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>