Remove calls in MODEL_GEOS blocks to Spc_Info which was deleted in 13.0
Spc_Info was used to get species long name which was then included as
the long name attribute in the GEOS-Chem internal state within GEOS.
However Spc_Info was retired in 13.0.0 and the equivalent information is
now stored in the State_Chm object. This information is therefore not
available when needed in GEOS since the internal state is set during
Set_Services which is prior to initializing GEOS-Chem state objects.

The species name is now used instead as a temporary work-around until
functionality is added in GEOS-Chem to get the long name from the
species database (now in yaml format) prior to instantiating the State_Chm
object.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>