CMakeList updates: print version and remove REAL*8 flags from main commands
We have updated the CMakeLists so that the version number (which is
stored in the PROJECT_VERSION variable) is copied to the include file
GeosCore/gc_classic_version.H.  This is done via template file
GeosCore/gc_classic_version.H.in.

NOTE: gc_classic_version.H is now generated by the CMake each time
GEOS-Chem is configured.  We have removed it and added it to the
.gitignore file so that it won't be tracked.  The template file
gc_classic_version.H.in has been added to the Git repository.

Also, if USE_REAL8 is selected, we add -DUSE_REAL8 to the compile
definitions list, but only add the extra compilation flags to
GeosCore/tpcore_window_mod.F90 (in the Transport library).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>